feat: Missing_val class
Class to handle missing value encodings have been created. The functions in the class decode the encoded missing values as np.nan and also check the data for encodings that have not been defined. To instantiate the class a attribute_mapping object need to be provided. Some doc strings are missing which will be added later.